--- tests/resources/caching/cache_root/database.json
@@ -1,1 +1,0 @@
-{"entries": [{"cache_id": 1, "identifier": "gAN9cQBYBAAAAGVkZ2VxAX1xAihYCQAAAG9wZXJhdGlvbnEDfXEEKFgEAAAAbmFtZXEFWAkAAAB0cmFuc2Zvcm1xBlgKAAAAcGFyYW1ldGVyc3EHfXEIWAgAAABmdW5jdGlvbnEJY2RpbGwuX2RpbGwKX2NyZWF0ZV9mdW5jdGlvbgpxCihjZGlsbC5fZGlsbApfbG9hZF90eXBlCnELWAgAAABDb2RlVHlwZXEMhXENUnEOKEsBSwBLAUsESxNDHHQAhwCHAWYCZAFkAoQIdAF8AIMBRACDAYMBUwBxD05oDihLAUsASwNLBEsTQyRnAHwAXRxcAn0BfQJ8AYgBawJyHIgAfAKDAW4CfAKRAnEEUwBxECkpWAIAAAAuMHERWAEAAABpcRJYBAAAAGVsZW1xE4dxFFg9AAAAL1VzZXJzL2F1NDc4MTA4L1Byb2plY3RzL2RhdGFzZXRvcHMvc3JjL2RhdGFzZXRvcHMvZGF0YXNldC5weXEVWAoAAAA8bGlzdGNvbXA+cRZN1wRDAgYAcRdYAgAAAGZucRhYAwAAAGlkeHEZhnEaKXRxG1JxHFhdAAAAX2RhdGFzZXRfZWxlbWVudF90cmFuc2Zvcm1pbmcuPGxvY2Fscz4ud3JhcHBlZC48bG9jYWxzPi5pdGVtX3RyYW5zZm9ybV9mbi48bG9jYWxzPi48bGlzdGNvbXA+cR2HcR5YBQAAAHR1cGxlcR9YCQAAAGVudW1lcmF0ZXEghnEhWAQAAABpdGVtcSKFcSNoFVgRAAAAaXRlbV90cmFuc2Zvcm1fZm5xJE3VBEMEAAECAXElaBhoGYZxJil0cSdScShjZGF0YXNldG9wcy5kYXRhc2V0Cl9fZGljdF9fCmgkTmNkaWxsLl9kaWxsCl9jcmVhdGVfY2VsbApxKWN0ZXN0aW5nX3V0aWxzCnJlYWRfYmluCnEqhXErUnEsaClLA4VxLVJxLoZxL31xME50cTFScTJzdVgGAAAAcGFyZW50cTN9cTRoAX1xNShoA31xNihoBWgGaAd9cTdoCWgKKGgoY2RhdGFzZXRvcHMuZGF0YXNldApfX2RpY3RfXwpoJE5oKWN0ZXN0aW5nX3V0aWxzCnJlYWRfdGV4dApxOIVxOVJxOmgpSwCFcTtScTyGcT19cT5OdHE/UnFAc3VoM31xQWgBfXFCKGgDfXFDKGgFaAZoB31xRGgJaAooaChjZGF0YXNldG9wcy5kYXRhc2V0Cl9fZGljdF9fCmgkTmgpY2RhdGFzZXRvcHMuZGF0YXNldApjb252ZXJ0MmltZwpxRYVxRlJxR2gpSwGFcUhScUmGcUp9cUtOdHFMUnFNc3VoM31xTmgBfXFPKGgDfXFQaAVYBAAAAGNvcHlxUXNoM31xUmgBXXFTKH1xVChoA31xVWgFWAMAAAB6aXBxVnNoM31xV2gBfXFYKFgEAAAAcm9vdHFZWFgAAAAvVXNlcnMvYXU0NzgxMDgvUHJvamVjdHMvZGF0YXNldG9wcy90ZXN0cy9yZXNvdXJjZXMvY2FjaGluZy9raXR0aV9kYXRhc2V0L3RyYWluaW5nL2NhbGlicVpoM051c3V9cVsoaAN9cVxoBWhWc2gzfXFdaAF9cV4oaFlYWgAAAC9Vc2Vycy9hdTQ3ODEwOC9Qcm9qZWN0cy9kYXRhc2V0b3BzL3Rlc3RzL3Jlc291cmNlcy9jYWNoaW5nL2tpdHRpX2RhdGFzZXQvdHJhaW5pbmcvaW1hZ2VfMnFfaDNOdXN1fXFgKGgDfXFhaAVoVnNoM31xYmgBfXFjKGhZWFoAAAAvVXNlcnMvYXU0NzgxMDgvUHJvamVjdHMvZGF0YXNldG9wcy90ZXN0cy9yZXNvdXJjZXMvY2FjaGluZy9raXR0aV9kYXRhc2V0L3RyYWluaW5nL2xhYmVsXzJxZGgzTnVzdX1xZShoA31xZmgFaFZzaDN9cWdoAX1xaChoWVhbAAAAL1VzZXJzL2F1NDc4MTA4L1Byb2plY3RzL2RhdGFzZXRvcHMvdGVzdHMvcmVzb3VyY2VzL2NhY2hpbmcva2l0dGlfZGF0YXNldC90cmFpbmluZy92ZWxvZHluZXFpaDNOdXN1ZXN1c3VzdXN1cy4="}], "last_id": 1}
