--- tests/resources/caching/kitti_dataset/training/label_2/000000.txt
@@ -1,1 +1,0 @@
-Pedestrian 0.00 0 -0.20 712.40 143.00 810.73 307.92 1.89 0.48 1.20 1.84 1.47 8.41 0.01
